Improve container resiliance to missing services/data. Fixes #3
