+++ .env.example
@@ -11,7 +11,7 @@ MONGO_INITDB_ROOT_PASSWORD=mongo_db_pass    # MongoDB password
 # Collector Configuration
 COLLECTOR_NR_USER=<nr-username>             # Network rail feed username
 COLLECTOR_NR_PASS=<nr-password>             # Network rail feed password
-COLLECTOR_ATTEMPTS=5                        # Number of STOMP connection attempts to make
+COLLECTOR_ATTEMPTS=3                        # Number of STOMP connection attempts to make
 COLLECTOR_PPM=True                          # Should PPM feed data be collected
 COLLECTOR_TD=True                           # Should TD feed data be collected
 COLLECTOR_TM=False                          # Should TM feed data be collected
